linux-firmware: add PRU Ethernet firmwares for AM57xx SoCs
Firmwares were added to be loaded onto each of the PRU Cores on
the AM572x SR2.0 and AM571x SR1.0 SoCs. These firmwares will
be supported by the TI PRUSS Ethernet driver. The firmwares
are modified for Linux to add the resource table needed by
the PRUSS remoteproc driver.

Signed-off-by: Andrew F. Davis <afd@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>
Acked-by: Dan Murphy <dmurphy@ti.com>
Acked-by: Roger Quadros <rogerq@ti.com>
Acked-by: Sekhar Nori <nsekhar@ti.com>