WHENCE: Update TI binary references to be WHENCE compliant
check_whence verifies that the binaries within the firmware
repo have WHENCE file references.  The format of the WHENCE
is specific.  Many TI binaries were tagged as not found in
the WHENCE file because the file tag was not properly defined.

Updating the WHENCE file to change "Files:" to "File:" and
adding this to each binary referenced.

Signed-off-by: Dan Murphy <dmurphy@ti.com>