linux-firmware: Update ICSSG Ethernet firmware on AM65
Update to release tag DEV.EMAC_LLD.01.00.03.15
This is actually an intermediate release before
Processor SDK release 06.xx.yy.zz.

It is actually DEV.PROCESSOR-SDK.05.03.00.05 with
some fixes in version string support.

There are a number of bug fixes, optimizations and
time stamping support for PTP when comparing to old
firmware.

It breaks old ABI, so device driver needs to be updated
as well.

Signed-off-by: Roger Quadros <rogerq@ti.com>
Acked-by: Dan Murphy <dmurphy@ti.com>